refactor: Move load functions into yaml/rule and yaml/channel
